brique-116/src/sira/requirements.txt
# Brique 116quinquies - Sira A/B Router Python Dependencies

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Optional: For async database operations
# asyncpg==0.29.0

# Optional: For data analysis
# pandas==2.1.4
# numpy==1.26.2

# Optional: For ML-based routing decisions
# scikit-learn==1.3.2

# Logging
python-json-logger==2.0.7

# Environment variables
python-dotenv==1.0.0
